Replace AssignSeatsToClientOrganizationCommand with ProviderBillingService, move to commercial
