+++ test/Api.Test/Billing/Controllers/ProviderClientsControllerTests.cs
@@ -7,6 +7,7 @@ using Bit.Core.AdminConsole.Entities.Provider;
 using Bit.Core.AdminConsole.Repositories;
 using Bit.Core.AdminConsole.Services;
 using Bit.Core.Billing.Commands;
+using Bit.Core.Billing.Services;
 using Bit.Core.Context;
 using Bit.Core.Entities;
 using Bit.Core.Models.Business;
@@ -327,7 +328,7 @@ public class ProviderClientsControllerTests
 
         var result = await sutProvider.Sut.UpdateAsync(providerId, providerOrganizationId, requestBody);
 
-        await sutProvider.GetDependency<IAssignSeatsToClientOrganizationCommand>().Received(1)
+        await sutProvider.GetDependency<IProviderBillingService>().Received(1)
             .AssignSeatsToClientOrganization(
                 provider,
                 organization,
